Initial Arduino support (#168)
* add UTL module

* split ufuncs into multiple files

* initial arduino suppor

* update tests

* add arduino examples

* update workflow

* try to fix arduino CI

* try to fix arduino CI (2)

* utl structured binding fix for android & emscripten

* try to fix gcc & clang workflow script

* update tests

* build utl tests by default

* utl structured binding fix for android & emscripten (2)

* specify python version on platformio ci

* fix mbed ci script error

* update android & wasm ci to set dockcross tag

* fix wasm ci script

* try to fix android ci script

* try to fix android and wasm error

* try to fix test for arm & riscv 32 bit

+++ include/nmtools/meta/bits/traits/is_num.hpp
@@ -15,8 +15,8 @@ namespace nmtools::meta
     template <> \
     struct is_floating_point<type> : true_type {};
 
-    NMTOOLS_IS_FLOATING_POINT_TRAIT(double)
-    NMTOOLS_IS_FLOATING_POINT_TRAIT(float)
+    NMTOOLS_IS_FLOATING_POINT_TRAIT(float32_t)
+    NMTOOLS_IS_FLOATING_POINT_TRAIT(float64_t)
 
 #undef NMTOOLS_IS_FLOATING_POINT_TRAIT
